requirements.txt
numpy
pydub
pyannote.audio
torch
ffmpeg-python

